requirements.txt
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
pydantic==2.5.2
pydantic_core==2.14.5
fastapi==0.100.1

h11==0.14.0
httptools==0.6.1
hume==0.4.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
urllib3==2.1.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
annotated-types==0.6.0
anyio==4.1.0
black==23.11.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0

idna==3.6
mypy-extensions==1.0.0

sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0

websockets==12.0
